\name{marys.elevslope}
\docType{data}
\alias{marys.elevslope}
\title{Elevation and Slope Data for Marys River, Oregon}
\description{
 The elevation and slope values for each segment (reach) in the
 Marys River stream network.
}
\usage{marys.elevslope}
\format{
 These data are derived from digital elevation models to describe
 topographic characteristics of the Marys River stream network:

 \tabular{ll}{
   arcid  \tab integer Arc/Info User-Id for each arc (segment)\cr
   elev   \tab elevation in meters\cr
   slope  \tab slope in percent
 }
}
\source{Derived from 30 meter and/or 10 meter Digital Elevation Models}
\keyword{datasets}
